refactor: move middleware to adapters
